# remove superfluous use of runtime.SetFinalizer on SQLiteRows